Move Offload Host handling to post-LTO stage.
Previously, we'd generate one host.o per CGU, overwriting each other.
We then would generate one more host.o during the final fat-lto run,
which was the one we ended up using. Move the host handling code, so
we stop generating the per-cgu host.o files that got overwritten anyway
