+++ compiler/rustc_codegen_llvm/src/back/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod archive;
 pub(crate) mod lto;
+pub(crate) mod offload;
 pub(crate) mod owned_target_machine;
 mod profiling;
 pub(crate) mod write;
